Restructure tomte
